requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7